Split frames::frame into a directory module
`frame.rs` had grown to hold four independent types. It becomes
`frame/`, with `object_id.rs`, `orbit_relative.rs`, and `body.rs` beside
a `mod.rs` that keeps `ReferenceFrame`, its constructors, and its
conversions. `mod.rs` re-exports each submodule, so every public path is
unchanged, and the tests move with the types they cover.
