+++ src/frames/frame/mod.rs
@@ -19,351 +19,19 @@
  */
 
 use std::fmt;
-use std::sync::Arc;
 
-use serde::de::{self, Visitor};
-use serde::{Deserialize, Deserializer, Serialize, Serializer};
+use serde::{Deserialize, Serialize};
 
 use crate::frames::CelestialFrame;
 use crate::utils::errors::BraheError;
 
-/// String-backed object identity (e.g. `"LRO"`, `"2024-123A"`).
-///
-/// Cheap to clone (`Arc<str>` internally). Serializes as a plain JSON
-/// string.
-///
-/// # Examples
-///
-/// ```rust
-/// use brahe::frames::ObjectId;
-///
-/// let id: ObjectId = "LRO".into();
-/// assert_eq!(id.to_string(), "LRO");
-/// assert_eq!(id, ObjectId::from("LRO".to_string()));
-/// ```
-#[derive(Clone, PartialEq, Eq, Hash, Debug)]
-pub struct ObjectId(Arc<str>);
+pub mod body;
+pub mod object_id;
+pub mod orbit_relative;
 
-impl fmt::Display for ObjectId {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        write!(f, "{}", self.0)
-    }
-}
-
-impl From<&str> for ObjectId {
-    fn from(value: &str) -> Self {
-        ObjectId(Arc::from(value))
-    }
-}
-
-impl From<String> for ObjectId {
-    fn from(value: String) -> Self {
-        ObjectId(Arc::from(value.as_str()))
-    }
-}
-
-impl Serialize for ObjectId {
-    fn serialize<S>(&self, serializer: S) -> Result<S::Ok, S::Error>
-    where
-        S: Serializer,
-    {
-        serializer.serialize_str(&self.0)
-    }
-}
-
-struct ObjectIdVisitor;
-
-impl Visitor<'_> for ObjectIdVisitor {
-    type Value = ObjectId;
-
-    fn expecting(&self, formatter: &mut fmt::Formatter) -> fmt::Result {
-        formatter.write_str("a string object identifier")
-    }
-
-    fn visit_str<E>(self, value: &str) -> Result<Self::Value, E>
-    where
-        E: de::Error,
-    {
-        Ok(ObjectId::from(value))
-    }
-}
-
-impl<'de> Deserialize<'de> for ObjectId {
-    fn deserialize<D>(deserializer: D) -> Result<Self, D::Error>
-    where
-        D: Deserializer<'de>,
-    {
-        deserializer.deserialize_str(ObjectIdVisitor)
-    }
-}
-
-/// Local orbital frame axes definitions.
-///
-/// `RTN` is the frame the SANA registries call `RSW`; brahe uses its
-/// existing RTN vocabulary (`state_eci_to_rtn`, `covariance_rtn`).
-///
-/// Every kind is a valid frame identity, which is what parsing a data file
-/// needs, but only `RTN` has an axes derivation today. A transform through
-/// any other kind errors until issue #452 adds the remaining derivations.
-///
-/// # Examples
-///
-/// ```rust
-/// use brahe::frames::OrbitRelativeFrameKind;
-///
-/// assert_eq!(OrbitRelativeFrameKind::RTN.to_string(), "RTN");
-/// ```
-#[derive(Debug, Clone, Copy, PartialEq, Eq, Serialize, Deserialize)]
-pub enum OrbitRelativeFrameKind {
-    /// Local-Vertical Local-Horizontal.
-    LVLH,
-    /// Radial / transverse (along-track) / normal (cross-track). SANA: RSW.
-    RTN,
-    /// Normal / tangential / cross-track.
-    NTW,
-    /// Tangential / normal / cross-track.
-    TNW,
-    /// Perifocal. SANA-registered only as an inertial-snapshot frame.
-    PQW,
-    /// Equinoctial. SANA-registered only as an inertial-snapshot frame.
-    EQW,
-    /// Topocentric south / east / zenith.
-    SEZ,
-    /// Velocity / normal / co-normal.
-    VNC,
-    /// Nadir / Sun / normal.
-    NSW,
-}
-
-impl fmt::Display for OrbitRelativeFrameKind {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        let token = match self {
-            Self::LVLH => "LVLH",
-            Self::RTN => "RTN",
-            Self::NTW => "NTW",
-            Self::TNW => "TNW",
-            Self::PQW => "PQW",
-            Self::EQW => "EQW",
-            Self::SEZ => "SEZ",
-            Self::VNC => "VNC",
-            Self::NSW => "NSW",
-        };
-        write!(f, "{}", token)
-    }
-}
-
-/// Rotating vs. quasi-inertial snapshot variant of a local orbital frame.
-///
-/// - **Rotating**: True local orbital frame, rotating with the orbit. It
-///   carries the orbital angular velocity, so a state transform through it
-///   picks up the corresponding velocity transport term.
-/// - **Inertial**: The same axes, taken as an instantaneous snapshot of the
-///   orbit state at the evaluation epoch and then treated as non-rotating.
-///   Its rate is zero, so a state transform through it applies no transport
-///   term. The axes still differ from epoch to epoch, since each evaluation
-///   takes a fresh snapshot.
-///
-/// # Examples
-///
-/// ```rust
-/// use brahe::frames::OrbitRelativeFrameVariant;
-///
-/// assert_eq!(OrbitRelativeFrameVariant::Rotating.to_string(), "rotating");
-/// ```
-#[derive(Debug, Clone, Copy, PartialEq, Eq, Serialize, Deserialize)]
-pub enum OrbitRelativeFrameVariant {
-    /// True local orbital frame, rotating with the orbit.
-    Rotating,
-    /// Quasi-inertial snapshot: the orbit-relative axes at the evaluation
-    /// epoch, with the frame's rate taken as zero.
-    Inertial,
-}
-
-impl fmt::Display for OrbitRelativeFrameVariant {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        match self {
-            Self::Rotating => write!(f, "rotating"),
-            Self::Inertial => write!(f, "inertial"),
-        }
-    }
-}
-
-/// A local orbital frame: a kind plus rotating/inertial-snapshot variant.
-///
-/// Represents a frame that rotates with or tracks the orbit, composed of:
-/// - A frame construction type (e.g., RTN, LVLH) defining the axes
-/// - A variant indicating whether the frame rotates with the orbit or is
-///   frozen
-///
-/// Fields are private: per the SANA registry, `EQW` and `PQW` exist only as
-/// inertial-snapshot frames, so construction goes through
-/// [`OrbitRelativeFrame::new`] to reject that combination rather than
-/// allowing it and erroring later.
-///
-/// # Examples
-///
-/// ```rust
-/// use brahe::frames::{OrbitRelativeFrame, OrbitRelativeFrameKind, OrbitRelativeFrameVariant};
-///
-/// let rtn = OrbitRelativeFrame::new(OrbitRelativeFrameKind::RTN, OrbitRelativeFrameVariant::Rotating);
-/// assert!(rtn.is_ok());
-/// assert_eq!(rtn.unwrap().to_string(), "RTN (rotating)");
-/// ```
-#[derive(Debug, Clone, Copy, PartialEq, Eq, Serialize, Deserialize)]
-pub struct OrbitRelativeFrame {
-    kind: OrbitRelativeFrameKind,
-    variant: OrbitRelativeFrameVariant,
-}
-
-impl OrbitRelativeFrame {
-    /// Constructs an orbit-relative frame, validating the kind/variant
-    /// combination.
-    ///
-    /// Per the SANA orbit-relative frame registry, `EQW` and `PQW` are
-    /// defined only as inertial-snapshot frames; they have no rotating
-    /// variant.
-    ///
-    /// # Arguments
-    /// * `kind` - The frame construction (axes definition)
-    /// * `variant` - Rotating (true local orbital frame) or inertial
-    ///   snapshot
-    ///
-    /// # Returns
-    /// * `Ok(OrbitRelativeFrame)`: If the combination is valid
-    /// * `Err(BraheError)`: If `kind` is `EQW` or `PQW` and `variant` is
-    ///   `Rotating`
-    ///
-    /// # Examples
-    ///
-    /// ```rust
-    /// use brahe::frames::{OrbitRelativeFrame, OrbitRelativeFrameKind, OrbitRelativeFrameVariant};
-    ///
-    /// let rtn = OrbitRelativeFrame::new(OrbitRelativeFrameKind::RTN, OrbitRelativeFrameVariant::Rotating);
-    /// assert!(rtn.is_ok());
-    ///
-    /// let eqw_rotating = OrbitRelativeFrame::new(OrbitRelativeFrameKind::EQW, OrbitRelativeFrameVariant::Rotating);
-    /// assert!(eqw_rotating.is_err());
-    /// ```
-    pub fn new(
-        kind: OrbitRelativeFrameKind,
-        variant: OrbitRelativeFrameVariant,
-    ) -> Result<Self, BraheError> {
-        if matches!(
-            kind,
-            OrbitRelativeFrameKind::EQW | OrbitRelativeFrameKind::PQW
-        ) && variant == OrbitRelativeFrameVariant::Rotating
-        {
-            return Err(BraheError::Error(format!(
-                "orbit-relative frame {} exists only as an inertial SANA frame and cannot be \
-                 constructed with the rotating variant",
-                kind
-            )));
-        }
-        Ok(Self { kind, variant })
-    }
-
-    /// Returns the frame construction (axes definition).
-    ///
-    /// # Returns
-    /// `OrbitRelativeFrameKind`: The frame construction
-    pub fn kind(&self) -> OrbitRelativeFrameKind {
-        self.kind
-    }
-
-    /// Returns the rotating/inertial-snapshot variant.
-    ///
-    /// # Returns
-    /// `OrbitRelativeFrameVariant`: Rotating or inertial
-    pub fn variant(&self) -> OrbitRelativeFrameVariant {
-        self.variant
-    }
-}
-
-impl fmt::Display for OrbitRelativeFrame {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        write!(f, "{} ({})", self.kind, self.variant)
-    }
-}
-
-/// An object-local spacecraft body frame with an optional instance
-/// designator.
-///
-/// The variants are the values of the SANA spacecraft body reference frame
-/// registry (<https://sanaregistry.org/r/spacecraft_body_reference_frames/>),
-/// covering spacecraft subsystems, sensors, and actuators. The optional
-/// `String` designator (e.g., `SCBody(Some("1"))`) is appended to the frame
-/// name in [`Display`](fmt::Display) output (e.g., `SC_BODY_1`).
-///
-/// # Examples
-///
-/// ```rust
-/// use brahe::frames::BodyFrame;
-///
-/// assert_eq!(BodyFrame::CSS(Some("1".to_string())).to_string(), "CSS_1");
-/// assert_eq!(BodyFrame::SCBody(None).to_string(), "SC_BODY");
-/// ```
-#[derive(Debug, Clone, PartialEq, Eq, Hash, Serialize, Deserialize)]
-pub enum BodyFrame {
-    /// Accelerometer frame.
-    ACC(Option<String>),
-    /// Actuator frame.
-    Actuator(Option<String>),
-    /// Autonomous star tracker frame.
-    AST(Option<String>),
-    /// Coarse sun sensor frame.
-    CSS(Option<String>),
-    /// Digital sun sensor frame.
-    DSS(Option<String>),
-    /// Earth sensor assembly frame.
-    ESA(Option<String>),
-    /// Gyroscope frame.
-    GyroFrame(Option<String>),
-    /// Inertial measurement unit frame.
-    IMUFrame(Option<String>),
-    /// Instrument frame.
-    Instrument(Option<String>),
-    /// Magnetic torque assembly frame.
-    MTA(Option<String>),
-    /// Reaction wheel frame.
-    RW(Option<String>),
-    /// Solar array frame.
-    SA(Option<String>),
-    /// Spacecraft body frame.
-    SCBody(Option<String>),
-    /// Generic sensor frame.
-    Sensor(Option<String>),
-    /// Star tracker frame.
-    StarTracker(Option<String>),
-    /// Three-axis magnetometer frame.
-    TAM(Option<String>),
-}
-
-impl fmt::Display for BodyFrame {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        let (family, suffix) = match self {
-            Self::ACC(s) => ("ACC", s),
-            Self::Actuator(s) => ("ACTUATOR", s),
-            Self::AST(s) => ("AST", s),
-            Self::CSS(s) => ("CSS", s),
-            Self::DSS(s) => ("DSS", s),
-            Self::ESA(s) => ("ESA", s),
-            Self::GyroFrame(s) => ("GYRO_FRAME", s),
-            Self::IMUFrame(s) => ("IMU_FRAME", s),
-            Self::Instrument(s) => ("INSTRUMENT", s),
-            Self::MTA(s) => ("MTA", s),
-            Self::RW(s) => ("RW", s),
-            Self::SA(s) => ("SA", s),
-            Self::SCBody(s) => ("SC_BODY", s),
-            Self::Sensor(s) => ("SENSOR", s),
-            Self::StarTracker(s) => ("STARTRACKER", s),
-            Self::TAM(s) => ("TAM", s),
-        };
-        match suffix {
-            Some(i) => write!(f, "{}_{}", family, i),
-            None => write!(f, "{}", family),
-        }
-    }
-}
+pub use body::*;
+pub use object_id::*;
+pub use orbit_relative::*;
 
 /// Unified frame identity spanning celestial, orbit-relative, and body
 /// frames.
@@ -1170,8 +838,8 @@ impl From<CelestialFrame> for ReferenceFrame {
 impl From<OrbitRelativeFrame> for ReferenceFrame {
     fn from(frame: OrbitRelativeFrame) -> Self {
         ReferenceFrame::OrbitRelative {
-            kind: frame.kind,
-            variant: frame.variant,
+            kind: frame.kind(),
+            variant: frame.variant(),
             object: None,
         }
     }
@@ -1271,149 +939,6 @@ mod tests {
             let s = serde_json::to_string(&f).unwrap();
             assert_eq!(serde_json::from_str::<ReferenceFrame>(&s).unwrap(), f);
         }
-    }
-
-    #[test]
-    #[parallel]
-    fn test_orbit_relative_kind_display_all_variants() {
-        let cases = [
-            (OrbitRelativeFrameKind::LVLH, "LVLH"),
-            (OrbitRelativeFrameKind::RTN, "RTN"),
-            (OrbitRelativeFrameKind::NTW, "NTW"),
-            (OrbitRelativeFrameKind::TNW, "TNW"),
-            (OrbitRelativeFrameKind::PQW, "PQW"),
-            (OrbitRelativeFrameKind::EQW, "EQW"),
-            (OrbitRelativeFrameKind::SEZ, "SEZ"),
-            (OrbitRelativeFrameKind::VNC, "VNC"),
-            (OrbitRelativeFrameKind::NSW, "NSW"),
-        ];
-        for (kind, expected) in cases {
-            assert_eq!(kind.to_string(), expected);
-        }
-    }
-
-    #[test]
-    #[parallel]
-    fn test_orbit_relative_variant_display() {
-        assert_eq!(OrbitRelativeFrameVariant::Rotating.to_string(), "rotating");
-        assert_eq!(OrbitRelativeFrameVariant::Inertial.to_string(), "inertial");
-    }
-
-    #[test]
-    #[parallel]
-    fn test_body_frame_display_all_variants() {
-        let cases = [
-            (BodyFrame::ACC(Some("1".to_string())), "ACC_1"),
-            (BodyFrame::Actuator(None), "ACTUATOR"),
-            (BodyFrame::AST(Some("1".to_string())), "AST_1"),
-            (BodyFrame::CSS(Some("2".to_string())), "CSS_2"),
-            (BodyFrame::DSS(Some("1".to_string())), "DSS_1"),
-            (BodyFrame::ESA(Some("1".to_string())), "ESA_1"),
-            (BodyFrame::GyroFrame(Some("1".to_string())), "GYRO_FRAME_1"),
-            (BodyFrame::IMUFrame(Some("2".to_string())), "IMU_FRAME_2"),
-            (BodyFrame::Instrument(Some("A".to_string())), "INSTRUMENT_A"),
-            (BodyFrame::MTA(Some("1".to_string())), "MTA_1"),
-            (BodyFrame::RW(Some("4".to_string())), "RW_4"),
-            (BodyFrame::SA(Some("1".to_string())), "SA_1"),
-            (BodyFrame::SCBody(None), "SC_BODY"),
-            (BodyFrame::Sensor(Some("10".to_string())), "SENSOR_10"),
-            (
-                BodyFrame::StarTracker(Some("2".to_string())),
-                "STARTRACKER_2",
-            ),
-            (BodyFrame::TAM(Some("1".to_string())), "TAM_1"),
-        ];
-        for (frame, expected) in cases {
-            assert_eq!(frame.to_string(), expected);
-        }
-    }
-
-    #[test]
-    #[parallel]
-    fn test_orbit_relative_frame_new_rejects_eqw_pqw_rotating() {
-        assert!(
-            OrbitRelativeFrame::new(
-                OrbitRelativeFrameKind::EQW,
-                OrbitRelativeFrameVariant::Rotating
-            )
-            .is_err()
-        );
-        assert!(
-            OrbitRelativeFrame::new(
-                OrbitRelativeFrameKind::PQW,
-                OrbitRelativeFrameVariant::Rotating
-            )
-            .is_err()
-        );
-        assert!(
-            OrbitRelativeFrame::new(
-                OrbitRelativeFrameKind::EQW,
-                OrbitRelativeFrameVariant::Inertial
-            )
-            .is_ok()
-        );
-        assert!(
-            OrbitRelativeFrame::new(
-                OrbitRelativeFrameKind::PQW,
-                OrbitRelativeFrameVariant::Inertial
-            )
-            .is_ok()
-        );
-    }
-
-    #[test]
-    #[parallel]
-    fn test_orbit_relative_frame_new_accepts_valid_combos() {
-        for kind in [
-            OrbitRelativeFrameKind::LVLH,
-            OrbitRelativeFrameKind::RTN,
-            OrbitRelativeFrameKind::NTW,
-            OrbitRelativeFrameKind::TNW,
-            OrbitRelativeFrameKind::SEZ,
-            OrbitRelativeFrameKind::VNC,
-            OrbitRelativeFrameKind::NSW,
-        ] {
-            assert!(OrbitRelativeFrame::new(kind, OrbitRelativeFrameVariant::Rotating).is_ok());
-            assert!(OrbitRelativeFrame::new(kind, OrbitRelativeFrameVariant::Inertial).is_ok());
-        }
-    }
-
-    #[test]
-    #[parallel]
-    fn test_orbit_relative_frame_kind_variant_accessors() {
-        let frame = OrbitRelativeFrame::new(
-            OrbitRelativeFrameKind::RTN,
-            OrbitRelativeFrameVariant::Rotating,
-        )
-        .unwrap();
-        assert_eq!(frame.kind(), OrbitRelativeFrameKind::RTN);
-        assert_eq!(frame.variant(), OrbitRelativeFrameVariant::Rotating);
-    }
-
-    #[test]
-    #[parallel]
-    fn test_orbit_relative_frame_display() {
-        let frame = OrbitRelativeFrame::new(
-            OrbitRelativeFrameKind::VNC,
-            OrbitRelativeFrameVariant::Rotating,
-        )
-        .unwrap();
-        assert_eq!(frame.to_string(), "VNC (rotating)");
-    }
-
-    #[test]
-    #[parallel]
-    fn test_object_id_from_string() {
-        let id = ObjectId::from(String::from("ISS"));
-        assert_eq!(id.to_string(), "ISS");
-        assert_eq!(id, ObjectId::from("ISS"));
-    }
-
-    #[test]
-    #[parallel]
-    fn test_object_id_deserialize_rejects_non_string() {
-        let err = serde_json::from_str::<ObjectId>("123").unwrap_err();
-        assert!(err.to_string().contains("a string object identifier"));
     }
 
     #[test]
